Split SimilarCustomersTable from ApplicationView
