+++ src/features/applicationView/ApplicationView.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import classNames from 'classnames';
 import { useTranslation } from 'react-i18next';
-import { Notification } from 'hds-react';
 import { PureQueryOptions } from 'apollo-client';
 
 import styles from './applicationView.module.scss';
@@ -9,8 +8,6 @@ import Card from '../../common/card/Card';
 import CardBody from '../../common/cardBody/CardBody';
 import ApplicationDetails, { ApplicationDetailsProps } from '../../common/applicationDetails/ApplicationDetails';
 import CardHeader from '../../common/cardHeader/CardHeader';
-import Table, { Column } from '../../common/table/Table';
-import CustomersTableTools, { CustomersTableToolsProps } from './customersTableTools/CustomersTableTools';
 import Text from '../../common/text/Text';
 import { formatDate } from '../../common/utils/format';
 import Chip from '../../common/chip/Chip';
@@ -18,11 +15,11 @@ import { APPLICATION_STATUS } from '../../common/utils/consonants';
 import CustomerProfileCard, { CustomerProfileCardProps } from '../../common/customerProfileCard/CustomerProfileCard';
 import OfferCard, { OfferCardProps } from './offerCard/OfferCard';
 import { CustomerGroup } from '../../@types/__generated__/globalTypes';
-import Pagination, { PaginationProps } from '../../common/pagination/Pagination';
 import PageTitle from '../../common/pageTitle/PageTitle';
 import PageContent from '../../common/pageContent/PageContent';
 import ActionHistoryCard from '../../common/actionHistoryCard/ActionHistoryCard';
-import { getCustomerGroupKey } from '../../common/utils/translations';
+import SimilarCustomersTableContainer from './SimilarCustomersTableContainer';
+import { INDIVIDUAL_APPLICATION_berthApplication as BERTH_APPLICATION } from './__generated__/INDIVIDUAL_APPLICATION';
 
 export enum SearchBy {
   FIRST_NAME = 'firstName',
@@ -40,78 +37,29 @@ export interface CustomerData {
   customerGroup: CustomerGroup | null;
 }
 
-type ColumnType = Column<CustomerData> & { accessor: keyof CustomerData };
-
 export interface ApplicationViewProps {
-  applicationId: string;
-  similarCustomersData: CustomerData[] | null;
-  customerProfile: CustomerProfileCardProps | null;
   applicationDetails: ApplicationDetailsProps;
+  berthApplication: BERTH_APPLICATION;
+  customerProfile: CustomerProfileCardProps | null;
   leaseDetails: OfferCardProps['leaseDetails'] | null;
-  customerTableTools: CustomersTableToolsProps<SearchBy>;
-  loadingCustomers?: boolean;
-  pagination: PaginationProps;
   refetchQueries: PureQueryOptions[] | string[];
-  handleLinkCustomer(customerId: string): void;
   handleDeleteLease(id: string): void;
-  onSortedColChange(sortBy: { id: string; desc?: boolean } | undefined): void;
 }
 
 const ApplicationView = ({
-  similarCustomersData,
-  customerProfile,
   applicationDetails,
+  berthApplication,
+  customerProfile,
   leaseDetails,
-  customerTableTools,
-  loadingCustomers,
-  pagination,
   refetchQueries,
-  handleLinkCustomer,
   handleDeleteLease,
-  onSortedColChange,
 }: ApplicationViewProps) => {
   const { t, i18n } = useTranslation();
-  const columns: ColumnType[] = [
-    {
-      Header: t('applicationView.customersTable.name') || '',
-      sortType: 'toString',
-      accessor: 'name',
-    },
-    {
-      Cell: ({ cell }) => {
-        const { value } = cell;
-        const customerGroupKey = getCustomerGroupKey(value);
-        return t(customerGroupKey);
-      },
-      Header: t('customerList.tableHeaders.group') || '',
-      disableSortBy: true,
-      accessor: 'customerGroup',
-    },
-    {
-      Header: t('applicationView.customersTable.municipality') || '',
-      disableSortBy: true,
-      accessor: 'city',
-    },
-    {
-      Header: t('applicationView.customersTable.address') || '',
-      disableSortBy: true,
-      accessor: 'address',
-    },
-    {
-      Cell: ({ cell }) => (
-        <div title={cell.value !== null ? cell.value : undefined} className={styles.berthsCell}>
-          {cell.value}
-        </div>
-      ),
-      Header: t('applicationView.customersTable.berths') || '',
-      disableSortBy: true,
-      accessor: 'berths',
-    },
-  ];
 
   return (
     <PageContent className={styles.applicationView}>
       <PageTitle title={t('applicationView.title')} />
+
       <div className={classNames(styles.fullWidth, styles.pageHeader)}>
         <Text as="h2" size="xl" weight="normalWeight">
           {applicationDetails.berthSwitch !== null
@@ -126,45 +74,22 @@ const ApplicationView = ({
         />
       </div>
 
-      {similarCustomersData && (
-        <>
-          <div className={styles.fullWidth}>
-            <Notification labelText={t('applicationView.noCustomerProfileNotification.label')} type="warning">
-              {t('applicationView.noCustomerProfileNotification.description')}
-            </Notification>
-          </div>
-          <Table
-            className={styles.fullWidth}
-            data={similarCustomersData}
-            loading={loadingCustomers}
-            columns={columns}
-            renderMainHeader={() => t('applicationView.customersTable.mainHeader')}
-            renderTableToolsTop={({ selectedRows }) => {
-              const onLinkCustomer = selectedRows.length ? () => handleLinkCustomer(selectedRows[0].id) : undefined;
-
-              return <CustomersTableTools {...customerTableTools} handleLinkCustomer={onLinkCustomer} />;
-            }}
-            renderTableToolsBottom={() => <Pagination {...pagination} className={styles.fullWidth} />}
-            renderEmptyStateRow={() => <div>{t('applicationView.customersTable.emptyState')}</div>}
-            onSortedColChange={onSortedColChange}
-            canSelectOneRow
-          />
-        </>
-      )}
-      {customerProfile && (
+      {customerProfile ? (
         <>
           <CustomerProfileCard {...customerProfile} />
           <ActionHistoryCard />
         </>
+      ) : (
+        <SimilarCustomersTableContainer berthApplication={berthApplication} />
       )}
-      {applicationDetails && (
-        <Card className={styles.fullWidth}>
-          <CardHeader title={t('applicationView.applicationDetails.title')} />
-          <CardBody>
-            <ApplicationDetails {...applicationDetails} handleDeleteLease={handleDeleteLease} queue={null} />
-          </CardBody>
-        </Card>
-      )}
+
+      <Card className={styles.fullWidth}>
+        <CardHeader title={t('applicationView.applicationDetails.title')} />
+        <CardBody>
+          <ApplicationDetails {...applicationDetails} handleDeleteLease={handleDeleteLease} queue={null} />
+        </CardBody>
+      </Card>
+
       {leaseDetails && (
         <OfferCard leaseDetails={leaseDetails} handleDeleteLease={handleDeleteLease} refetchQueries={refetchQueries} />
       )}
